safari-600.7-branch/LayoutTests/platform/mac/http/tests/inspector/console-websocket-error-expected.txt

CONSOLE MESSAGE: line 32: testDNSLookup: Test started.
CONSOLE MESSAGE: WebSocket network error: The operation couldn’t be completed. (kCFErrorDomainCFNetwork error 2.)
CONSOLE MESSAGE: line 43: testDNSLookup: onclose is called.
CONSOLE MESSAGE: line 50: testSSLCertificate: Test started.
CONSOLE MESSAGE: WebSocket network error: OSStatus Error -9807: Invalid certificate chain 
CONSOLE MESSAGE: line 64: testSSLCertificate: onclose is called.
WebSocket's network errors should be logged to console.

testDNSLookup: Test started. console-websocket-error.html:32
WebSocket network error: The operation couldn’t be completed. (kCFErrorDomainCFNetwork error 2.) http://127.0.0.1:8000/inspector/console-websocket-error.html
testDNSLookup: onclose is called. console-websocket-error.html:43
testSSLCertificate: Test started. console-websocket-error.html:50
WebSocket network error: OSStatus Error -9807: Invalid certificate chain  http://127.0.0.1:8000/inspector/console-websocket-error.html
testSSLCertificate: onclose is called. console-websocket-error.html:64
